pyproject.toml
[tool.poetry]
name = "agr_literature_service"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.27.1"
pytest = "^7.0.0"
black = "^22.1.0"
mypy = "^0.931"
flake8 = "^4.0.1"
SQLAlchemy = "^1.4.31"
alembic = "^1.7.6"
fastapi = "^0.73.0"
PyYAML = "^6.0"
fastapi-health = "^0.4.0"
uvicorn = "^0.17.4"
SQLAlchemy-Continuum = "^1.3.12"
pydot = "^1.4.2"
boto = "^2.49.0"
bcrypt = "^3.2.0"
python-jose = "^3.3.0"
python-multipart = "^0.0.5"
psycopg2-binary = "^2.9.3"
asyncio = "^3.4.3"
asyncpg = "^0.25.0"
passlib = "^1.7.4"
python-dotenv = "^0.19.2"
SQLAlchemy-Utils = "^0.38.2"
pytz = "^2021.3"
databases = "^0.5.5"
fastapi-auth0 = "^0.3.0"
isort = "^5.10.1"
Sphinx = "^4.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.isort]
profile = "hug"
src_paths = ["agr_literature_service"]
py_version=39
skip = [".gitignore", ".dockerignore", "__init__.py", "agr_literature_service/api/main.py"]
skip_glob = ["docs/*"]
multi_line_output = 3

[tool.pytest.ini_options]
addopts = "--cov --cov-fail-under=80 -vv --cov-report html --ignore non_pr_tests/"
